Created SqlServerManageAdder to add MessageBus listerners if ManagementChannelName specified.
Added some additional "Adder" to add the FileJanitor and Database message handlers to the installation when deploying with an InitializationStrategySqlServer.  Added tests around adding in general and slightly tweaked in the interface to provide setup factory settings as some of the defaults are useful.  Fixed the InitializationSelfHost ExeName property to include path like ScheduledTask one.  Added support for SQL Server Named Instances.